+++ test_defopt.py
@@ -1,5 +1,6 @@
 import builtins
 import contextlib
+import functools
 import inspect
 import multiprocessing as mp
 import re
@@ -21,8 +22,8 @@ from docutils.utils import SystemMessage
 import defopt
 from defopt import __version__, _options
 from examples import (
-    annotations, booleans, choices, exceptions, lists, parsers, short,
-    starargs, styles)
+    annotations, booleans, choices, exceptions, lists, parsers, partials,
+    short, starargs, styles)
 
 
 Choice = Enum('Choice', [('one', 1), ('two', 2), ('%', 0.01)])
@@ -1172,6 +1173,30 @@ class TestAlias(unittest.TestCase):
         self.assertEqual(defopt.run(d['foo'], argv=['-b1']), [1])
 
 
+class TestFunctoolsPartial(unittest.TestCase):
+    # does not have a default for `bar`
+    @staticmethod
+    def foo(*, bar: int) -> None:
+        """The foo tool
+
+        :param int bar: the bar option"""
+        return bar
+
+    def test_partial_top_level(self):
+        actual = defopt.run(functools.partial(self.foo, bar=4), argv=[])
+        self.assertEqual(actual, 4)
+
+    def test_partial_sub_command(self):
+        funcs = {'foo': functools.partial(self.foo, bar=4)}
+        actual = defopt.run(funcs, argv=['foo'])
+        self.assertEqual(actual, 4)
+
+    def test_partial_sub_commands(self):
+        funcs = {'bar': [functools.partial(self.foo, bar=4)]}
+        actual = defopt.run(funcs, argv=['bar', 'foo'])
+        self.assertEqual(actual, 4)
+
+
 class TestHelp(unittest.TestCase):
     def test_type(self):
         def foo(bar):
@@ -1335,6 +1360,39 @@ class TestHelp(unittest.TestCase):
 
         self.assert_in_help('summary-of-foo', [foo, bar], '')
         self.assert_not_in_help('FOO', [foo, bar], '')
+
+    def test_functools_partial_make_param_default(self):
+        def foo(*, bar):
+            """The foo tool
+
+            :param int bar: the bar option"""
+        func = functools.partial(foo, bar=4)
+        self.assert_in_help('The foo tool', func, 't')
+        self.assert_in_help('(type: int)', func, 't')
+        self.assert_in_help('(default: 4)', func, 'd')
+        self.assert_in_help('(type: int, default: 4)', func, 'dt')
+
+    def test_functools_partial_override_param_default(self):
+        def foo(*, bar=5):
+            """The foo tool
+
+            :param int bar: the bar option"""
+        func = functools.partial(foo, bar=4)
+        self.assert_in_help('The foo tool', func, 't')
+        self.assert_in_help('(type: int)', func, 't')
+        self.assert_in_help('(default: 4)', func, 'd')
+        self.assert_in_help('(type: int, default: 4)', func, 'dt')
+
+    def test_functools_partial_command_list(self):
+        def foo():
+            """Foo"""
+        def bar():
+            """Bar"""
+        func_foo = functools.partial(foo)
+        func_bar = functools.partial(bar)
+        funcs = [func_foo, func_bar]
+        self.assert_in_help('Foo', funcs, 't')
+        self.assert_in_help('Bar', funcs, 't')
 
     def assert_in_help(self, s, funcs, flags):
         self.assertIn(s, self._get_help(funcs, flags))
@@ -1510,6 +1568,12 @@ class TestExamples(unittest.TestCase):
             self._run_example(parsers, ['junk'])
         self.assertIn(b'datetime', error.exception.output)
         self.assertIn(b'junk', error.exception.output)
+
+    def test_partials(self):
+        output = self._run_example(partials, ['foo'])
+        self.assertEqual(output, b'5\n')
+        output = self._run_example(partials, ['sub', 'bar'])
+        self.assertEqual(output, b'6\n')
 
     def test_short(self):
         with self._assert_stdout('hello!\n'):